requirement.txt
langchain-core==0.1.53
langchain-openai==0.0.8
langchain-community==0.0.32
langchainhub==0.1.15
langchain-groq
langchain-objectbox

python-dotenv==1.0.1
streamlit==1.35.0
fastapi==0.110.0
uvicorn==0.29.0
sse-starlette==1.6.5
bs4==0.0.2
beautifulsoup4==4.12.3
pypdf==3.17.4
PyPDF2==3.0.1
chromadb==0.4.24
faiss-cpu==1.7.4
sentence-transformers==2.6.1
wikipedia==1.4.0
arxiv==1.4.7
cassio==0.1.4
langserve==0.0.51
fastapi
uvicorn
langserve
pydantic<2.0
fastapi==0.110.0
langchain
langchain-openai
langchain-community
langserve
uvicorn
sse_starlette
python-dotenv
streamlit>=1.36
python-dotenv>=1.0
langchain>=0.2.12
langchain-community>=0.2.10
langchain-groq>=0.1.3
sentence-transformers>=3.0
chromadb>=0.5.3
pypdf>=4.2.0
uvicorn>=0.30  # if/when you add FastAPI
fastapi>=0.114
